requirements.txt
boto3>=1.26.0
pydantic>=2.0.0
hydra-core>=1.3.0
transformers>=4.30.0
torch>=2.0.0
pytest>=7.0.0
python-dotenv>=1.0.0
clearml>=1.14.0  # for experiment tracking
psutil>=5.9.0  # for memory tracking
typer>=0.9.0  # for CLI interface
